+++ src/Components/LicenseAlert/LicenseAlert.tsx
@@ -11,27 +11,27 @@ const LicenseAlert: React.FC<LicenseAlertProps> = ({ alertMessage }) => (
     message={
       alertMessage || (
         <span>
-          Please review the{" "}
+          ReductStore Core is licensed under the{" "}
           <strong>
             <a
               href="https://github.com/reductstore/reductstore/blob/main/LICENSE"
               target="_blank"
               rel="noopener noreferrer"
             >
-              Business Source License (BUSL)
+              Apache License 2.0
             </a>
           </strong>{" "}
-          for ReductStore and consult our{" "}
+          and commercial licensing options for{" "}
           <strong>
             <a
               href="https://www.reduct.store/pricing"
               target="_blank"
               rel="noopener noreferrer"
             >
-              pricing page
+              ReductStore Pro
             </a>
           </strong>{" "}
-          for more information on licensing options.
+          are available on our pricing page.
         </span>
       )
     }